Rename useProfileSectionTypes as default export
